requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.3
pydantic-settings==2.2.1
sqlalchemy==2.0.28
aiosqlite==0.20.0
redis==5.0.2
rq==1.16.1
httpx==0.27.0
python-dotenv==1.0.1
openai==1.13.3
dashscope==1.19.0
moviepy==1.0.3
ffmpeg-python==0.2.0
Pillow==10.2.0
aiofiles==23.2.1
anyio==4.3.0
aiohttp==3.9.5
